don't we need an extra dense softmax layer?
Yes, it would be a good idea to add a dense layer with the softmax activation function at the end of the model to obtain the probability distribution over the possible values of the input. This would allow you to use a cross-entropy loss function when training the model, which is typically used for classification tasks.

Here's an example of how you could modify the WaveNetModel to include a final dense layer with the softmax activation